requirements.txt
requests==2.31.0
anthropic>=0.40.0
python-dotenv==1.0.0
cryptography==42.0.0